requirements.txt
fastapi==0.78.0
pydantic==1.10.26
python-dotenv==1.2.1
requests==2.32.5
requests-cache==1.3.1
SQLAlchemy==2.0.49
uvicorn==0.17.6